Calculate poverty severity using microdata

Non-vectorized poverty severity microdata function. Use the vectorized function pipmd_pov_severity

Usage

pipmd_pov_severity_nv(
  pipster_object = NULL,
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  povline = fmean(welfare, w = weight) * times_mean,
  times_mean = 1
)

Arguments

pipster_object

pipster object created using create_pipster_object

welfare

numeric: A vector of income or consumption values

weight

numeric: A vector of population weights. If NULL, a vector of 1s is used to give equal weight to each observation.

povline

numeric: Poverty line in international dollars, same units as welfare.

times_mean

numeric factor that multiplies the mean to create a relative poverty line. Default is 1

Value

numeric: Poverty severity